requirements.txt
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
node==1.2
odict==1.9.0
plumber==1.7
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0
